SkinDiseaseModel/requirements.txt

streamlit
torch
numpy
opencv-python
pillow
reportlab
